Fixes https://github.com/sindresorhus/is-plain-obj/issues/12

This improves the performance.

Benchmarks on Node 18.3.0 on my machine, before the change:

```
undefined                                         : 491.366ms
0                                                 : 545.934ms
0n                                                : 8.169s
''                                                : 730.961ms
true                                              : 689.757ms
Symbol()                                          : 8.272s
[Function (anonymous)]                            : 687.628ms
[Function: namedFunc]                             : 679.116ms
null                                              : 508.197ms
{}                                                : 1.493s
Object [Math] {}                                  : 3.589s
Set(0) {}                                         : 3.841s
ArrayBuffer { [Uint8Contents]: <>, byteLength: 0 }: 2.408s
Promise { undefined }                             : 2.397s
[Object: null prototype] {}                       : 2.138s
Locale [Intl.Locale] {}                           : 3.197s
{ prop: true }                                    : 1.480s
Class {}                                          : 3.277s
[]                                                : 726.733ms
/regexp/                                          : 758.549ms
function Error() { [native code] }                : 721.147ms
2022-06-13T21:32:59.396Z                          : 692.619ms
[Arguments] {}                                    : 652.13ms
{}                                                : 13.990s
```

After the change:

```
undefined                                         : 161.508ms
0                                                 : 122.998ms
0n                                                : 121.546ms
''                                                : 121.174ms
true                                              : 121.634ms
Symbol()                                          : 121.894ms
[Function (anonymous)]                            : 121.306ms
[Function: namedFunc]                             : 121.261ms
null                                              : 143.305ms
{}                                                : 1.081s
Object [Math] {}                                  : 1.141s
Set(0) {}                                         : 980.147ms
ArrayBuffer { [Uint8Contents]: <>, byteLength: 0 }: 969.084ms
Promise { undefined }                             : 968.839ms
[Object: null prototype] {}                       : 2.520s
Locale [Intl.Locale] {}                           : 985.919ms
{ prop: true }                                    : 2.397s
Class {}                                          : 979.786ms
[]                                                : 966.707ms
/regexp/                                          : 973.039ms
function Error() { [native code] }                : 966.822ms
2022-06-13T21:31:59.071Z                          : 972.304ms
[Arguments] {}                                    : 3.602s
{}                                                : 10.165s
```

In a nutshell:
  - Some less-used native types (`symbol`, `bigint`) are 60 times faster
  - Most native types (`undefined`, `null`, `number`, `string`, `boolean`) and functions are 5 times faster
  - Plain objects are 50% times faster. This is important since they are the most likely argument.
  - Class instances and many core object types (`Set`, `Promise`, `ArrayBuffer`, `Math`, `Intl.*`, `arguments`) are 2-4 times faster
  - `Proxy` is 30% faster
  - `Object.create(null)` is 10% slower
  - `Array`, `RegExp`, `Error`, `Date` are 30% slower. This is the main drawback. However, this seems worthwhile based on the other improvements.
  - `new Object()` is twice slower. It is rarely used though.