@sindresorhus 2022-06-14
Can you fix the conflict?