@ehmicky 2022-06-14
For completeness, those are benchmarks before/after both this PR and the cross-realm PR (#14). Before:

```
undefined                                         : 545.663ms
0                                                 : 591.312ms
0n                                                : 9.436s
''                                                : 838.412ms
true                                              : 786.609ms
Symbol()                                          : 9.429s
[Function (anonymous)]                            : 752.251ms
[Function: namedFunc]                             : 773.338ms
null                                              : 564.996ms
{}                                                : 1.661s
Object [Math] {}                                  : 3.891s
Set(0) {}                                         : 4.195s
ArrayBuffer { [Uint8Contents]: <>, byteLength: 0 }: 2.678s
Promise { undefined }                             : 3.159s
[Object: null prototype] {}                       : 2.586s
Locale [Intl.Locale] {}                           : 3.797s
{ prop: true }                                    : 1.725s
Class {}                                          : 3.597s
[]                                                : 846.749ms
/regexp/                                          : 848.499ms
function Error() { [native code] }                : 844.467ms
2022-06-14T16:56:35.326Z                          : 797.05ms
[Arguments] {}                                    : 748.679ms
{}                                                : 16.165s
```

After:

```
undefined                                         : 165.378ms
0                                                 : 123.947ms
0n                                                : 121.877ms
''                                                : 122.252ms
true                                              : 123.71ms
Symbol()                                          : 123.806ms
[Function (anonymous)]                            : 122.966ms
[Function: namedFunc]                             : 122.122ms
null                                              : 142.183ms
{}                                                : 1.125s
Object [Math] {}                                  : 1.143s
Set(0) {}                                         : 1.718s
ArrayBuffer { [Uint8Contents]: <>, byteLength: 0 }: 1.705s
Promise { undefined }                             : 1.703s
[Object: null prototype] {}                       : 2.567s
Locale [Intl.Locale] {}                           : 1.792s
{ prop: true }                                    : 3.600s
Class {}                                          : 1.800s
[]                                                : 1.801s
/regexp/                                          : 1.798s
function Error() { [native code] }                : 1.790s
2022-06-14T16:55:42.385Z                          : 1.786s
[Arguments] {}                                    : 3.901s
{}                                                : 10.792s
```