@sindresorhus 2022-06-15
Nice work :)