@sindresorhus 2022-06-15
Would you be interested in applying your changes to https://github.com/sindresorhus/is/blob/dc99f7cd4a8ad309d4c945b21816132dbc09ef92/source/index.ts#L287-L297 too?