@ehmicky 2022-06-15
Sure! Done at https://github.com/sindresorhus/is/pull/169